deploy app to kind cluster
The app defined by proxy/Dockerfile can now be deployed to the kind cluster.

First, the cluster has now a private registry that is connected to the network and exposed to the host machine as well.

We can build the image locally and push it to this registry. The image is then accessible inside the cluster.

The pod is allowed to talk to the cluster because we create a ClusterRole and a ClusterRoleBinding.

The ingress-controller app now start nginx and remains in a loop.

We can run the script to update the nginx.conf as well.

The next step is making the nginx server see the conf produced by the app